+++ CHANGELOG.md
@@ -16,9 +16,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Added
+
+- Append log data when sending exception to Bugsnag
+
 ### Changed
 
-- Updated Bugsnag library and exception handling to v2 API
+- Update Bugsnag library and exception handling to v2 API
 - [launcher] Improve robustness
 
 ## [1.0.2] - 2018-04-09